Implement StimulusTests!
This adds a route to the app that routes to a statically-configurable
controller, but only in the test env. The controller can have its layout
set before each visit, along with the HTML to render on the page.

When visited, the given HTML (or block that's evaluated in view_context)
is rendered into the layout (or nothing), and the importmap entry point
can be configured too (best used when the layout is nil). This allows
any kind of HTML to be rendered such that a Stimulus Controller will get
executed on the page, and can be asserted upon.